Get rid of dead code guarded when USE_BOOST not defined
