+++ PsimagLite/src/Ainur/Ainur.h
@@ -1,10 +1,6 @@
 #ifndef _AINUR_H_
 #define _AINUR_H_
 
-#ifdef USE_BOOST
 #include "AinurSpirit.h"
-#else
-#include "AinurEmpty.h"
-#endif
 
 #endif // _AINUR_H_
